Allow users to expand a step to show its props
